Title: 사각형 채우기 3, Time: 139ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240910/사각형 채우기 3/rectangle-fill-3.java
@@ -2,7 +2,7 @@ import java.util.*;
 
 public class Main {
     public static int n;
-    public static int[] dp = new int[1001];
+    public static long[] dp = new long[1001];
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         n = sc.nextInt();
